Consolidate geocoding code
- CLGeocoder is deprecated in OS 26.0. So encapsulate it here in preparation for refactoring.
