+++ Sources/Site/Music/Atlas.swift
@@ -13,10 +13,6 @@ extension Logger {
   fileprivate static let atlas = Logger(category: "atlas")
 }
 
-enum GeocodeError: Error {
-  case noPlacemark
-}
-
 protocol Geocodable {
   func geocode() async throws -> CLPlacemark
 }
